move field-definition code to new module
